Restructure + LINQ use in CommonController
