+++ Ranker/Repository/Core/IRankerRepository.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Ranker
+{
+    public interface IRankerRepository
+    {
+        IRanksRepository Ranks { get; }
+
+        IRolesRepository Roles { get; }
+    }
+}
